Can build docs locally without failing.
