+++ qiskit_metal/toolbox_metal/__init__.py
@@ -44,6 +44,9 @@ Submodules
     import_export
     math_and_overrides
     parsing
+    layer_stack_handler
+    bounds_for_path_and_poly_tables
+    determine_larger_box
 
 """
 
@@ -52,7 +55,7 @@ from .parsing import is_variable_name
 from .parsing import is_numeric_possible
 from .parsing import parse_units
 from .layer_stack_handler import LayerStackHandler
-from .bounds_for_path_and_poly_tables import BoundsForPathAndPolyTables
+from .bounds_for_path_and_poly_tables import BoundsForPathAndPolyTables, determine_larger_box
 
 from .. import config
 if config.is_building_docs():
